requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pymongo==4.13.2
motor==3.7.1
python-dotenv==1.1.1
email_validator==2.2.0
pydantic-settings==2.10.1
python-jose==3.5.0
cryptography==45.0.5
bcrypt==3.2.2
passlib==1.7.4
# for OAuth2PasswordRequestForm (for handling form data in login route)
python-multipart==0.0.20
Faker==37.5.3
flake8==7.3.0
black==25.11.0
isort==7.0.0
pre_commit==4.4.0
